+++ cmd/tools/pmanager/main.go
@@ -8,6 +8,7 @@ import (
 	"os"
 
 	"github.com/number571/go-peer/pkg/crypto/keybuilder"
+	"github.com/number571/go-peer/pkg/crypto/symmetric"
 )
 
 func main() {
@@ -16,7 +17,7 @@ func main() {
 	flag.Parse()
 
 	keyBuilder := keybuilder.NewKeyBuilder(1<<(*workParam), []byte(*saltParam))
-	gotPassword := keyBuilder.Build(readUntilEOL())
+	gotPassword := keyBuilder.Build(readUntilEOL(), symmetric.CAESKeySize)
 
 	fmt.Println(base64.URLEncoding.EncodeToString(gotPassword))
 }