+++ pkg/crypto/keybuilder/keybuilder_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"testing"
 
+	"github.com/number571/go-peer/pkg/crypto/symmetric"
 	"github.com/number571/go-peer/pkg/encoding"
 	testutils "github.com/number571/go-peer/test/utils"
 )
@@ -20,13 +21,13 @@ func TestKeyBuilder(t *testing.T) {
 		salt = []byte("it's a salt!")
 	)
 
-	hash := NewKeyBuilder(1<<testutils.TCWorkSize, salt).Build(pasw)
+	hash := NewKeyBuilder(1<<testutils.TCWorkSize, salt).Build(pasw, symmetric.CAESKeySize)
 	if encoding.HexEncode(hash) != tcHash {
 		t.Error("hash is correct?")
 		return
 	}
 
-	if !bytes.Equal(hash, NewKeyBuilder(1<<testutils.TCWorkSize, salt).Build(pasw)) {
+	if !bytes.Equal(hash, NewKeyBuilder(1<<testutils.TCWorkSize, salt).Build(pasw, symmetric.CAESKeySize)) {
 		t.Error("hash is not determined")
 		return
 	}